[main-theme] Re-implement and simplify Chord algorithm (#8)
### Motivation

There are so many functions in `ChordProtocol`, but that is not usable
to implement dummy. It is needed to be simplified. Also, from DDD, the
entity has an identifier and is mutable. Thus I hope to use the mutable
state for local data. As we know, the remote hosts have their own data,
but there is no requirement to sync the datas in Chord DHT, so it will
be implemented as immutable values.

<!--Explain why you're making this change and what problem you're trying
to solve.-->

### Modifications

- Implement `AbstractChordNode` and their mutable `MutableChordNode` and
immutable `ChordNode` subclasses.
- Add `ChordScope` to convert values or nodes to `Id`.
- Move mutation logic to `MutableChordNode`.
- Simplify `ChordProtocol` by removing the redundant functions.

<!--List the modifications you've made in detail.-->

### Result

- Addition and implementation of the above modifications.

<!--Closes #. (If this resolves the issue.)-->
<!--Describe the consequences that a user will face after this PR is
merged.-->