[main-theme] Refactor to be testable for each events (#13)
### Motivation

<!--Explain why you're making this change and what problem you're trying
to solve.-->

For the testable application. we can do the following things.

- Remove mutable data classes.
- Separate logics from `Service` to each classes.

These leads the logics to be more testable without many dependencies.
Also, this approach can be found in the theory.

#### Domain Event from DDD

In DDD, the domain event is defined as followings. In short, the
**domain events** are events related to the changes in the domain
models(aggregates), which domain experts care about.

> A domain event is a full-fledged part of the domain model, a
representation of something that happened in the domain. Ignore
irrelevant domain activity while making explicit the events that the
domain experts want to track or be notified of, or which are associated
with state changes in the other model objects.

The Chord algorithm shows the solution of the P2P hash table. I think
the domain model is Chord Node, and the domain event is methods of
managing the key-values. So I designed each logics. which are setting
values / getting values / finding the nodes, as a domain events.

#### UseCase from Clean Architecture 

Also, there are some inferences from the Clean architecture.

> The software in this layer contains application specific business
rules. It encapsulates and implements all of the use cases of the
system. These use cases orchestrate the flow of data to and from the
entities, and direct those entities to use their enterprise wide
business rules to achieve the goals of the use case.

With the same argument, the use case of Chord algorithm is managing
key-values. So they can be seen as use cases.

### Modifications

<!--List the modifications you've made in detail.-->

- Address the commits from the previous PR.
  - Add `InvalidChordNode`.
  - Remove the mutable `ChordNode`.
- Separate the domain events from `Service`.
- Add interfaces to handle the local datas.

### Result

We can have separated logics to each classes, and they represent the
domain event or use case. Also we can get each test codes without each
classes.

<!--Closes #. (If this resolves the issue.)-->
<!--Describe the consequences that a user will face after this PR is
merged.-->

### Next works

- Remove too many type arguments if it is not needed.
- Implement Repository to remove values so that the new node has its
data.

### Reference

-
https://medium.com/@chaojie.xiao/domain-driven-design-practice-domain-events-15b38f3c58fc
-
https://blog.cleancoder.com/uncle-bob/2012/08/13/the-clean-architecture.html