[archive] Add introduction to optimal product quantization (#17)
### Motivation

For the simplicity search in a large-size database, we need to much time
to get the exact data. This will discourage our product to users because
of the processing time. There is a technique of finding the most similar
vector in the database. That is the optimal product quantization. Before
entering to the main content, the introduction is needed. This PR will
add the introduction of the main story.

<!--Explain why you're making this change and what problem you're trying
to solve.-->

### Modifications

The vector database is an issue for the similarity search of a large
data. The background of the database will help us to understand the
database too.

<!--List the modifications you've made in detail.-->

### Result

The introduction of the content is added.

- What is the similarity search
- What is the vector quantization
- Introduction to the k-means

<!--Closes #. (If this resolves the issue.)-->
<!--Describe the consequences that a user will face after this PR is
merged.-->

### Next work

- Write the main content of the optimized product quantization