[archive] Add definition and algorithm of OPQ and its result (#24)
### Motivation

<!--Explain why you're making this change and what problem you're trying
to solve.-->

This PR will cover the content of OPQ, and its result. This will handle
some mathematical point of view, but we can see the result of this OPQ
with the RECALL graph. Please note that the recall is the value of the
number of true positive(TP) over the sum of true positive(TP) and false
negative(FN), i.e. $Recall = TP / (TP + FN)$.

### Modifications

<!--List the modifications you've made in detail.-->

- PQ
  - Formal definition
- OPQ
  - Formal definition
  - Algorithm
  - Initial $R$
  - Experiment results
- Fix of un-rendered latex

### Result

We can introduce the OPQ solution of ANN problem. There are some other
points to implement this, like algorithms for building codebook, but we
can cover the concept of OPQ.

<!--Closes #. (If this resolves the issue.)-->
<!--Describe the consequences that a user will face after this PR is
merged.-->